Struct qemu_rust::qapi_connection::QApiConnection

pub struct QApiConnection {
    // some fields omitted
}

The connection to the qemu QMP service.

Methods

impl QApiConnection

fn new(sock: TcpStream) -> QApiConnection

Create a new connection to Qemu. On my localhost I usually start qemu in another terminal like this: qemu-system-i386 -qmp tcp:localhost:4444,server,nowait

fn run_command<E, T: Decodable>(&self, cmd: E) -> Result<T, String> where E: QemuCmd<T>

Run a Qemu command and get a parsed json respose back